requirements.txt
fastapi
uvicorn
pydantic
sqlalchemy
black
bcrypt
passlib
python-jose
python-multipart